+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Conditions, weapons and armors 
 - Character inventory
+- Character summary mode
 
 ## [2.4.1] - 2019-04-09
 ### Fixed